+++ modules/api/src/main/java/org/fossnova/http2/protocol/HeadersFrame.java
@@ -24,10 +24,10 @@ package org.fossnova.http2.protocol;
  * @author <a href="mailto:opalka.richard@gmail.com">Richard Opalka</a>
  */
 public interface HeadersFrame extends Frame {
-    int FLAG_END_STREAM = 0x1;
-    int FLAG_END_HEADERS = 0x4;
-    int FLAG_PADDED = 0x8;
-    int FLAG_PRIORITY = 0x20;
+    byte FLAG_END_STREAM = 0x1;
+    byte FLAG_END_HEADERS = 0x4;
+    byte FLAG_PADDED = 0x8;
+    byte FLAG_PRIORITY = 0x20;
 
     int getPadLength();
     int getStreamDependency();
