+++ modules/api/src/main/java/org/fossnova/http2/protocol/PushPromiseFrame.java
@@ -24,8 +24,8 @@ package org.fossnova.http2.protocol;
  * @author <a href="mailto:opalka.richard@gmail.com">Richard Opalka</a>
  */
 public interface PushPromiseFrame extends Frame {
-    int FLAG_END_HEADERS = 0x4;
-    int FLAG_PADDED = 0x8;
+    byte FLAG_END_HEADERS = 0x4;
+    byte FLAG_PADDED = 0x8;
 
     int getPadLength();
     int getPromisedStreamId();
